Completed operator overload code in Signed32 class. Needs more testing though.
Implementation of operator overloading is complete in Signed32. Just
need to do some more testing in both to assure the behavior is correct,
and then add the new ToFloat() method to both Unsigned32 and Signed32. I
may also add more documentation (at least an overview, if not a full
class write up) before I push the new version to the hub.