b1.2.2.20171227 release
A minor note about the version number: Based on my numbering scheme, the
last update should probably have been b1.2.1.20171217, so I'm skipping
over this version number and jumping to the correct order. To be honest,
b1.1.0.20170724 should probably have been b1.2.0.20170724 too, but oh
well...

pif_LongInt:
- Fixed a bug that would result in empty arguments (except in
constructors) would result in a runtime error.
- Fixed a bug in the operator-() method on the Signed32 class that would
fail to return the proper result when doing negation (that is, when the
operator-() method had no arguments).
- Changed the behavior of right bitshifting methods (BitshiftRight(),
operator>>(), and operator>>=()) on the Signed32 class. Previously, the
most significant bits were filled with 0's, while common behavior is to
fill them with 1's if the number is negative so as to keep right
bitshifts the same as integer division by two. The BitshiftRightRotate()
method is not affected by this change.
- Implemented FindFirstSet(), FindLastSet(), CountLeadingZeros(), and
CountTrailingZeros() on both Signed32 and Unsigned32.

pif_Arithmetic:
- Updated the specification on the FindFirstSet() and FindLastSet()
methods.
- Added CountLeadingZeros() and CountTrailingZeros() methods.